feat: 接口 新增模版，新增供应商

+++ src/main/java/com/zte/msg/pushcenter/utils/Constants.java
@@ -9,5 +9,5 @@ package com.zte.msg.pushcenter.utils;
  */
 public class Constants {
 
-    public static long DATA_NOT_DELETED = 0;
+    public static int DATA_NOT_DELETED = 0;
 }
